Removes the staggered grid layout
This Re-aligns the guide section and all sections with a staggered grid layout. Attempts to change the icons were made and I spent many hours trying to learn how to add custom icons that would fit in these cards. The current framework seems almost hostile to making custom icon changes, so this would involve making changes to the components themselves and is a much larger task than anticipated. Very strong possibility to continue this as a separate card in the future.