Option 2 Updates Automatically
This version updates all dependencies automatically, so we would never have to manually do this again and the site will always be updated. However, this could result in unexpected issues if an update breaks something, in which case it would have to be fixed or manually reverted to an earlier build. Astro-icon has been left untouched as any updates to this currently results in a null reference and causes issues with the current site.